zmij::stdarch_x86

Function _mm256_movepi32_mask
·

pub fn _mm256_movepi32_mask(a: __m256i) -> u8

Set each bit of mask register k based on the most significant bit of the corresponding packed 32-bit integer in a.

Intel’s Documentation