zmij::stdarch_x86

Function _mm512_mask_cmul_round_pch
·

pub fn _mm512_mask_cmul_round_pch<const ROUNDING: i32>(
    src: __m512h,
    k: u16,
    a: __m512h,
    b: __m512h,
) -> __m512h

Multiply packed complex numbers in a by the complex conjugates of packed complex numbers in b, and store the results in dst using writemask k (the element is copied from src when corresponding mask bit is not set). Each complex number is composed of two adjacent half-precision (16-bit) floating-point elements, which defines the complex number complex = vec.fp16[0] + i * vec.fp16[1], or the complex conjugate conjugate = vec.fp16[0] - i * vec.fp16[1].

Rounding is done according to the rounding parameter, which can be one of:

_MM_FROUND_TO_NEAREST_INT | _MM_FROUND_NO_EXC : round to nearest and suppress exceptions

_MM_FROUND_TO_NEG_INF | _MM_FROUND_NO_EXC : round down and suppress exceptions

_MM_FROUND_TO_POS_INF | _MM_FROUND_NO_EXC : round up and suppress exceptions

_MM_FROUND_TO_ZERO | _MM_FROUND_NO_EXC : truncate and suppress exceptions

_MM_FROUND_CUR_DIRECTION : use MXCSR.RC - see _MM_SET_ROUNDING_MODE

Intel’s documentation