zmij::stdarch_x86

Function _mm512_maskz_permutex2var_ps
·

pub fn _mm512_maskz_permutex2var_ps(
    k: u16,
    a: __m512,
    idx: __m512i,
    b: __m512,
) -> __m512

Shuffle single-precision (32-bit) floating-point elements in a and b across lanes using the corresponding selector and index in idx, and store the results in dst using zeromask k (elements are zeroed out when the corresponding mask bit is not set).

Intel’s documentation